updates to deal with more general declaration types

+++ rote/languages/langc/examples/guard_ex1.c
@@ -1,6 +1,6 @@
 int foo() {
   int *x;
-  int *y;
+  double *y;
   x = malloc(10);
   y = malloc(12);
   // if (x == null) exit(1);   <--- insert this
